Broke out builtin functions in main into separate utility library
