+++ gvutils/displayVersion.go
@@ -1,9 +1,9 @@
-package main
+package gvutils
 
 import (
 	"fmt"
 )
 
-func displayVersion(version, buildTime string) {
+func DisplayVersion(version, buildTime string) {
 	fmt.Printf("Current Version is %s, built at %s.\n", version, buildTime)
 }
