+++ gvutils/flagparse.go
@@ -1,16 +1,17 @@
-package main
+package gvutils
 
 import (
 	"flag"
 )
 
-func flagParser() (action string) {
+func FlagParser() (action string) {
 	verS := flag.Bool("v", false, "displays the current version and build time")
 	verL := flag.Bool("version", false, "displays the current version and build time")
 	flag.Parse()
 
-	if (verS || verL) == true {
+	if *verS || *verL == true {
 		action = "version"
 		return
 	}
+	return
 }
